Set EAPOL-Key Key Length field to 0 for group message 1/2 in RSN
P802.11i/D3.0 described the Key Length as having value 16 for the group
key handshake. However, this was changed to 0 in the published IEEE Std
802.11i-2004 amendment (and still remains 0 in the current standard IEEE
Std 802.11-2016). We need to maintain the non-zero value for WPA (v1)
cases, but the RSN case can be changed to 0 to be closer to the current
standard.

Change-Id: Iaeea9354da2c530dbf836c0a4d4af22af304c369
Signed-off-by: Jouni Malinen <j@w1.fi>
Signed-off-by: Marco Zanin (B--B) <mrczn.bb@gmail.com>