Fix PTK rekeying to generate a new ANonce
The Authenticator state machine path for PTK rekeying ended up bypassing
the AUTHENTICATION2 state where a new ANonce is generated when going
directly to the PTKSTART state since there is no need to try to
determine the PMK again in such a case. This is far from ideal since the
new PTK would depend on a new nonce only from the supplicant.

Fix this by generating a new ANonce when moving to the PTKSTART state
for the purpose of starting new 4-way handshake to rekey PTK.

Change-Id: I3edfba00ac8e89ec4c93585f3fa206935c87a742
Signed-off-by: Jouni Malinen <j@w1.fi>
Signed-off-by: Marco Zanin (B--B) <mrczn.bb@gmail.com>